refactor(ui): split PathDiscoveryCard into L2/L3/helpers modules
Extract sub-components from PathDiscoveryCard.tsx (970 -> 585 lines,
under the 800-line red-flag budget).

- pathDiscoveryHelpers.ts: formatRtt, getMaxRtt, getRttBarColor,
  getSourceColor pure helpers shared across the displays
- PathDiscoveryCardL3.tsx: L3_PATH_DISPLAY traceroute hop list
- PathDiscoveryCardL2.tsx: L2_PATH_DISPLAY plus L2_HOP_DETAIL and
  PORT_DETAILS subviews for the LLDP/CDP/SNMP switch chain

Pure relocation, no behavior change. Public PathDiscoveryCard export
and TraceHopMessage type are unchanged.
